+++ scripts/mk-shadowdark.js
@@ -1,5 +1,5 @@
-import "./summary-bar.js";
-import "./sheet-style-editor.js";
+import "./summary-bar/summary-bar.js";
+import "./sheet-style-editor/sheet-style-editor.js";
 
 (() => {
   const MODULE_ID = "mk-shadowdark";